// aws_lambda_vul/app/src/types.ts
export enum Exploitation {
    Active = 'active',
    Poc = 'poc',
    None = 'none'
  }
  
export enum Exposure {
    Open = 'open',
    Controlled = 'controlled',
    Small = 'small'
  }
  
export enum Utility {
    SuperEffective = 'superEffective',
    Efficient = 'efficient',
    Laborious = 'laborious'
  }
  
export enum Automatable {
    Yes = 'yes',
    No = 'no'
  }
  
export enum ValueDensity {
    Diffuse = 'diffuse',
    Concentrated = 'concentrated'
  }
  
export enum HumanImpact {
    VeryHigh = 'veryHigh',
    High = 'high',
    Medium = 'medium',
    Low = 'low'
  }